Siguiente Actualizacion, los enemigos se mueven de forma random y tienen vidas

+++ Assets/_Assets/Scripts/Enemigos/EnemyWeapons.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class EnemyWeapons : MonoBehaviour {
     [SerializeField] Transform bala;
-    [SerializeField] int cadencia =2;
+    [SerializeField] float cadencia =2;
     float tiempo = 0;
     [SerializeField] int MaxDist = 30;
     [SerializeField] int MinDist = 10;
